\name{display.alts}
\alias{display.alts}
\title{ Display Alt Codes }
\description{
A utility function to display a selection of the Alt codes available from a PC keyboard.
}
\usage{
display.alts()
}
\details{
Based on a function shared on S-News.
}
\author{ Unknown }
\seealso{ \code{\link{display.ascii.o}}, \code{\link{display.ascii.d}} }
\examples{
display.alts()
}
\keyword{ misc }

