\name{display.ascii.d}
\alias{display.ascii.d}
\title{ Display the ASCII Character Set by Decimal Number }
\description{
A utility function to display the decimal numbers corresponding to the Windows Latin 1 Font.
}
\usage{
display.ascii.d()
}
\details{
Based on a function shared on S-News for octal numbers.
}
\author{ Robert G. Garrett }
\seealso{ \code{\link{display.ascii.o}}, \code{\link{display.alts}} }
\examples{
display.ascii.d()
}
\keyword{ misc }

