\name{display.ascii.o}
\alias{display.ascii.o}
\title{ Display the Windows Latin 1 Font Octal Table }
\description{
A utility function to display the octal numbers corresponding to the Windows Latin 1 Font.
}
\usage{
display.ascii.o()
}
\details{
The ASCII octal \sQuote{escape codes} are used to insert special characters in text strings for axis labelling, and titles etc., in graphical displays.  For example the escape string \code{\265} results in the Greek letter mu being displayed.

Based on a function shared on S-News.
}
\author{ Unknown }
\examples{
display.ascii.o()
}
\keyword{ misc }

