\name{display.lty}
\alias{display.lty}
\title{ Display Available Line Styles and Colour Codes }
\description{
Displays the line styles and colours corresponding to \code{lty = 1} to \code{9} and \code{colr = 1} to \code{9}, respectively.
}
\usage{
display.lty()
}
\details{
Several rgr functions that plot boxes or polygons may have their default infill colour of grey, \code{colr = 8}, changed to an alternate colour, \code{colr = 1} to \code{7} or \code{9}.  For no infill colour, set \code{colr} = 0.
}
\author{ Robert G. Garrett }
\keyword{ misc }

