\name{display.marks}
\alias{display.marks}
\title{ Display Available Plotting Marks }
\description{
Displays the available plotting marks.  Where specified, the \sQuote{rgr} functions use a plus sign, \code{pch = 3}, as the plotting symbol, alternate plotting marks may be selected from this display.  For example, \code{pch = 1} results in an open circle, the R default, and \code{pch = 4} results in a cross, \sQuote{x}.  For additional symbols available only in R (\code{pch = 19:25}) see \code{\link{points}}.
}
\usage{
display.marks()
}
\author{ Various, see Note }
\note{
Function to display \code{pch} codes based on a script originally shared on S-News by Bill Venables, 1996/07, and modified by Shawn Boles, 1996/07/31.
}
\keyword{ hplot }
\keyword{ misc }
