\name{display.rainbow}
\alias{display.rainbow}
\title{ Display the Colours of the Rainbow(36) Pallette }
\description{
Displays the available colours in the rainbow(36) pallette to support the selection of alternate colour schemes.
}
\usage{
display.rainbow()
}
\author{ Robert G. Garrett }
\keyword{ color }
\keyword{ misc }
