\name{fix.test}
\docType{data}
\alias{fix.test}
\title{ Test Data for Function ltdl.fix.df }
\description{
A set of test data to demonstrate how negative values are changed to half their positive value.  Optionally numeric coded values representing missing data and/or zero values may be replaced by \code{NA}s.
}
\usage{fix.test}
\format{ A dataframe containing 15 rows and 4 columns (2 factors and 2 numeric). }
\keyword{ datasets }








