\name{map.eda7}
\alias{map.eda7}
\title{ Plot a Symbol Map of Data Based on the Tukey Boxplot }
\description{
Displays a simple map where the data are represented at their spatial locations by symbols using Tukey boxplot-based symbology. Tukey boxplots divide data into 7 groups, the middle 50\%, and three lower and higher groupings, see Details below. The computation of the fences used to subdivide the data may be carried out following a logarithmic transformation of the data. The colours of the symbols may be optionally changed.
}
\usage{
map.eda7(xx, yy, zz, sfact = 1, logz = FALSE, xlab = "Easting", 
	ylab = "Northing", zlab = deparse(substitute(zz)), main = "", 
	ifgrey = FALSE, symcolr = NULL, tol = 0.04, ...)
}
\arguments{
  \item{xx}{ name of the x-axis spatial coordinate, the eastings. }
  \item{yy}{ name of the y-axis spatial coordinate, the northings. }
  \item{zz}{ name of the variable to be plotted. }
  \item{sfact}{ controls the absolute size of the plotted symbols, by default \code{sfact = 1}.  Increasing sfact results in larger symbols. }
  \item{xlab}{ a title for the x-axis, defaults to \dQuote{Easting}. }
  \item{ylab}{ a title for the y-axis, defaults to \dQuote{Northing}. }
  \item{zlab}{ by default, \code{zlab = deparse(substitute(zz))}, a map title is generated by appending the input variable name text string to \dQuote{EDA Tukey Boxplot Map for}. Alternative titles may be generated, see Details below. }
  \item{main}{ an alternative map title, see Details below. }
  \item{logz}{ if it is required to undertake the Tukey Boxplot computations after a logarithmic data transform, set \code{log = TRUE}. }
  \item{ifgrey}{ set \code{ifgrey = TRUE} if a grey-scale map is required, see Details below. }
  \item{symcolr}{ the default is a colour map and default colours are provided, deeper blues for lower values, green for the middle 50\% of the data, and oranges and reds for higher values.  A set of alternate symbol colours can be provided by defining \code{symcolr}, see Details below. }
  \item{tol}{ a parameter used to ensure the area included within the neatline around the map is larger than the distribution of the points so that the plotted symbols fall within the neatline.  By default \code{tol = 0.04}, if more clearance is required increase the value of \code{tol}. }
  \item{\dots}{ further arguments to be passed to methods.  For example, if it is required to make the map title smaller, add \code{cex.main = 0.9} to reduce the font size by 10\%. }
}
\details{
Tukey boxplots divide data into 7 groups, the middle 50\%, and three lower and high groupings: within the whisker, near outliers and far outliers, respectively.  Symbols for values below the first quartile (Q1) are plotted as increasingly larger circles, while symbols for values above the third quartile are plotted as increasingly larger squares.  For the higher groupings, the whisker contains values \code{>Q3 and <(Q3 + 1.5 * HW)}, where \code{HW = (Q3 - Q1)}, the interquartile range; near outliers lie between \code{(Q3 + 1.5 * HW)} and \code{(Q3 + 3 * HW)}; and far outliers have values \code{>(Q3 + 3 * HW)}.  For the lower groupings the group boundaries, fences, fall similarly spaced below Q1.  The computation of the fences used to subdivide the data may be carried out following a logarithmic transformation of the data, set \code{logz = TRUE}.

A summary table of the values of the symbol intervals, the number of values plotting as each symbol, and symbol shapes, sizes and colours is displayed on the current device.

If \code{zlab} and \code{main} are undefined a default a map title is generated by appending the input variable name text string to \dQuote{EDA Tukey Boxplot-Based Map for}.  If no map title is required set \code{zlab = ""}, and if some user defined map title is required it should be defined in \code{main}, e.g. \code{main = "Map Title Text"}.

If the grey-scale option is chosen the symbols are plotted 100\% black for the far outliers, 85\% black for the near outliers, 70\% black for values within the whiskers, and 60\% black for values falling within the middle 50\% of the data.

The default colours, \code{symcolr = c(25, 22, 20, 13, 6, 4, 1)}, are selected from the \code{rainbow(36)} pallette, and alternate colour schemes need to be selected from the same palette.  See \code{\link{display.rainbow}} for the available colours.  It is essential that 7 colours be provided, e.g., \code{symcolr = c(27, 24, 22, 12, 5, 3, 36)}, if exactly 7 are not provided the default colours will be displayed.
}
\note{
Any less than detection limit values represented by negative values, or zeros or other numeric codes representing blanks in the data vector, must be removed prior to executing this function, see \code{\link{ltdl.fix.df}}.

Any \code{NA}s in the data vector are removed prior to displaying the plot.

In some R installations the generation of multi-panel displays and the use of function eqscplot from package MASS causes warning messages related to graphics parameters to be displayed on the current device. These may be suppressed by entering \code{options(warn = -1)} on the R command line, or that line may be included in a \sQuote{first} function prepared by the user that loads the rgr package, etc.
}
\author{ Robert G. Garrett }
\seealso{ \code{\link{display.rainbow}}, \code{\link{ltdl.fix.df}}, \code{\link{remove.na}} }  
\examples{ 
## Make test data available
data(kola.o)
attach(kola.o)

## Plot a default symbol map
map.eda7(UTME, UTMN, Cu)

## Plot with logarithmically scaled boxplot fences and more
## appropriate axis labelling
map.eda7(UTME/1000, UTMN/1000, Cu, logz = TRUE,
	xlab = "Kola Project UTM Eastings (km)",
	ylab = "Kola Project UTM Northings (km)") 

## Plot a grey-scale equivalent of the above map
map.eda7(UTME/1000, UTMN/1000, Cu, logz = TRUE, ifgrey = TRUE,
	xlab = "Kola Project UTM Eastings (km)",
	ylab = "Kola Project UTM Northings (km)") 

## Plot the same map with an alternate colour scheme
map.eda7(UTME/1000, UTMN/1000, Cu, logz = TRUE, 
	xlab = "Kola Project UTM Eastings (km)",
	ylab = "Kola Project UTM Northings (km)",
	symcolr = c(27, 24, 22, 12, 5, 3, 36)) 

## Detach test data
detach(kola.o) 
}
\keyword{ hplot }


