\name{nockolds}
\alias{nockolds}
\alias{Nockolds}
\docType{data}
\title{ The Nockolds average igneous plutonic rock compositions }
\description{
A data frame of average major and minor element compositions for igneous plutonic rocks published by Nockolds (1954).
}
\usage{data(nockolds)}
\format{
  A data frame with 16 average igneous plutonic rock compositions for 12 parts:
  \describe{
    \item{Row_ID}{ four character abbreviation for lithology. } 
    \item{Si}{ silicon (cation \%). }
    \item{Al}{ aluminium (cation\%). }
    \item{Fe3}{ ferric iron (cation \%). }
    \item{Fe2}{ ferrous iron (cation \%). }
    \item{Mg}{ magnesium (cation \%). }
    \item{Ca}{ calcium (cation \%). }
    \item{Na}{ sodium (cation \%). }
    \item{K}{ potassium (cation \%). }
    \item{Ti}{ titanium (cation \%). }
    \item{Mn}{ manganese (cation \%). }
    \item{P}{ phosphorus (cation \%). }
    \item{OH}{ hydroxyl (anion \%). }
  }
  Row_ID abbreviations for igneous plutonic rocks:
  \describe{
    \item{ALKG}{ Alkali granite }
    \item{GRNT}{ Granite }
    \item{QZMZ}{ Quartz monzonite }
    \item{GRDR}{ Granodiorite }
    \item{QRZD}{ Quartz diorite }
    \item{ALKS}{ Alkali syenite }
    \item{SENT}{ Syenite }
    \item{MNZN}{ Monzonite }
    \item{MZDT}{ Monzodiorite }
    \item{DORT}{ Diorite }
    \item{GBBR}{ Gabbro }
    \item{PRDT}{ Peridotite }
    \item{ANRS}{ Anorthosite }
    \item{NPLS}{ Nepheline Syenite }
    \item{ESXT}{ Essexite }
    \item{IJLT}{ Ijolite }
  }
}
\source{
Nockolds, S.R., 1954. Average chemical compositions of some igneous rocks.  Geological Society of Americe, Bulletin 65, pp. 1007-1032. 
}
\references{
Garrett, R.G., Reimann, C., Hron, K., Kynclova, P. and Filzmoser, P., 2017. Finally, a correlation coefficient that tells the geochemical truth. Explore - Assoc. Applied Geochemists Newsletter, 176:1-15. 
}
\keyword{datasets}
