\name{syms.pfunc}
\alias{syms.pfunc}
\title{ Function to Demonstrate the Effect of Different Values of p }
\description{
This function displays a plot demonstrating the effect of varying the value of p, for a range of p values from 0.2 to 5, on the 0 to 1 normalized values of a variable in order to compute corresponding circular symbol diameters. 
}
\usage{
syms.pfunc()
}
\author{ Robert G. Garrett }
\keyword{ hplot }


