\name{display.marks}
\alias{display.marks}
\title{ Display Available Plotting Marks }
\description{
Displays the available plotting marks.  By default the rgr functions use a plus sign, \code{pch = 3}, as the plotting symbol, alternate plotting marks may be selected from this display.  For example, \code{pch = 1} results in an open circle, and \code{pch = 4} results in a cross/x.
}
\usage{
display.marks()
}
\author{ Robert G. Garrett }
\note{
Function to display \code{pch} codes based on a script originally shared on S-News by Bill Venables, 1996/07, and modified by Shawn Boles, 1996/07/31.
}
\keyword{ misc }
