\name{rgr-package}
\alias{rgr-package}
\alias{rgr}
\docType{package}
\title{
The GSC (Geological Survey of Canada) Applied Geochemistry EDA
Package
}
\description{
R functions for Exploratory Data Analysis with applied geochemical data. 
}
\details{
The functions in this package are used to support the display and analysis of applied geochemical survey data, particularly in the context of estimating the ranges of background variation due to natural phenomena and the identification of outliers that may be due to natural processes or anthropogenic contamination.  This version, 1.1.1, includes a number of multivariate tools that have proven useful in recognizing structure in data sets and identifying multivariate outliers for both open and closed, compositional geochemical, data sets.  Where several background populations can be recognized and defined an allocation procedure is provided to classify individuals into the background populations or place them in an \sQuote{unknown} category as outliers for further inspection.
}
\author{
Robert G. Garrett

Maintainer: Robert G. Garrett <garrett@NRCan.gc.ca>
}
\references{
Reimann, C., Filzmoser, P. and Garrett, R.G., 2005. Background and threshold: critical comparison of methods of determination. Science of the Total Environment, 346(1/3):1-16.

Reimann, C. and Garrett, R.G., 2005. Geochemical background - Concept and reality. Science of the Total Environment, 350(1/3):12-27.

Reimann, C., Filzmoser, P., Garrett, R. and Dutter, R., 2008. Statistical Data Analysis Explained: Applied Environmental Statistics with R. John Wiley & Sons, Ltd., 362 p.

Venables, W.N. and Ripley, B.D., 2001. Modern Applied Statistics with S-Plus, 3rd Edition, Springer, 501 p.
}
\keyword{ package }
